First verified working version tested against adiabatic hydrodynamics example using vac4.52 with cd4 and rk4. vactest directory includes the test files for vac4.52. Main problem was setting boundary conditions for the derived quantities.

+++ dev/sac2.5d-cuda/iotypes.h
@@ -8,7 +8,7 @@
 #ifdef USE_REAL
 DEFINE_PRECISION(float)
 #else
-DEFINE_PRECISION(double)
+DEFINE_PRECISION(float)
 #endif
